Fix negative priority Priority is an integer between -128 and +127.

+++ mcabber/src/jabglue.c
@@ -309,7 +309,7 @@ static xmlnode presnew(enum imstatus st, const char *recipient,
   prio = settings_opt_get_int("priority");
   if (prio) {
     char strprio[8];
-    snprintf(strprio, 8, "%u", prio);
+    snprintf(strprio, 8, "%d", (int)prio);
     xmlnode_insert_cdata(xmlnode_insert_tag(x, "priority"),
                          strprio, (unsigned) -1);
   }
